refactor(cli): extract upgrade cluster into commands/upgrade.ts (5.0-alpha)
Third cut of the 5.0-alpha cli.ts god-module split. Moves cmdUpgrade,
selfUpgrade, and the two npm-permission-error helpers (isNpmPermissionError,
npmPermissionRemediation) out of cli.ts into src/commands/upgrade.ts.

- All four are exported: cmdUpgrade (COMMANDS entry), selfUpgrade (still
  called by cli.ts's opt-in auto-update flow), and the two error helpers
  (re-exported from cli.ts so upgrade-self-error.test.ts's `from "./cli.js"`
  import keeps working).
- cli.ts imports cmdUpgrade + selfUpgrade back and re-exports the two helpers;
  the COMMANDS dispatch table is unchanged.
- lock.js imports stay in cli.ts (cmdInit still uses readkitMeta /
  updateSkillsLock / updateCliLock).
- Dynamic imports re-pathed ./ -> ../ (triage-gate, utils/exec,
  utils/redactSecrets).

cli.ts: 4544 -> 4417 lines (-127; -1913 cumulative across PRs 1-3).
Typecheck, eslint (0 errors), full test suite (2580, incl.
upgrade-self-error), public-surface snapshot, and prettier all green.
